+++ hygen/_templates/layer/new/layer.ejs.t
@@ -12,7 +12,7 @@ export const <%= h.changeCase.constantCase(name) %>_ID = '<%= h.changeCase.camel
 export default function <%= h.changeCase.pascalCase(name) %>() {
   const { <%= h.changeCase.camelCase(name) %> } = useSelector((state) => state.carto.layers);
   const source = useSelector((state) => selectSourceById(state, <%= h.changeCase.camelCase(name) %>?.source));
-  const cartoLayerProps = useCartoLayerProps(source);
+  const cartoLayerProps = useCartoLayerProps({ source });
 
   if (<%= h.changeCase.camelCase(name) %> && source) {
     return new CartoLayer({