+++ scripts/bootstrap.sh
@@ -10,7 +10,7 @@ set_template()
   cd -
 }
 
-set_template template-sample-app
-set_template template-skeleton
-set_template template-skeleton-cn
+set_template template-base-2
+set_template template-sample-app-2
+set_template template-base-3