+++ template-base-2/template/src/routes.js
@@ -14,6 +14,7 @@ export const ROUTE_PATHS = {
   DEFAULT: '/',
   OAUTH: '/oauthCallback',
   NOT_FOUND: '404',
+  STORES: '/stores',
   // [hygen] Add path routes
 };